requirements.txt
aiogram==3.12.0
aiogram_dialog==2.2.0b3
cryptography==42.0.8
ujson==5.10.0
openpyxl==3.1.2
redis==5.0.8
python-dotenv==1.0.1
aioclock==0.3.0
easyocr==1.7.1
pillow==10.2.0
pytz==2024.1
termcolor==2.4.0
playwright==1.46.0
pydantic==2.5.3